// src/types/navigation.ts
export type RootStackParamList = {
  Onboarding: undefined;
  Subdomain: undefined;
  Login: undefined;
  Drawer: undefined;
  MachineDetail: { machineId: number; machineName: string };
  Notifications: undefined;
};

export type DrawerParamList = {
  Dashboard: undefined;
  Machines: undefined;
  Reports: undefined;
  Users: undefined;
  Settings: undefined;
  Notifications: undefined;
  Logout: undefined;
}; 